requirements.txt
google-generativeai
pdfplumber
pypdf
flask
